make: Enable ASIO driver use
Using the ASIO driver with the plugin will allow audio playback
from compatible sound cards with much lower latency.

Find the ASIO SDK with GLOB and include this in the building process.
Set JUCE_ASIO=1 to enable ASIO use.
This requires the Steinberg ASIO SDK to be downloaded and added
to the directory.

Add ASIO SDK to the gitignore file as it is now required in the
directory.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>