editor: Draw frequency response
Declare monoChain as public.

override and declare paramter changed functions so that the monochain
in plugin editor can be updated asyncronously to match any updates in
the processor audio change.

Multiply the magnitudes of all the filters in the chain that are not
bypassed.
Draw the combined response curve of each filter in the chain.

Signed-off-by Jon Friedman <jfriedman97@gmail.com>