processor: Add filter bypass booleans
Add bypass booleans to APVTS layout.
Add filter bypass booleans to getChainSettings function.
These are stored as floats and should be set to true if they are greater
than 0.5f.
Add filter bypass states to the ChainSettings struct.
Bypass lowcut, highcut and peak filters in the audio chain if they are
bypassed in the chain settings.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>